requirements.txt
GitPython==2.1.8
plumbum
requests